Add build system and wrapper for adaptive coordinate
Because the adaptive coordinate core is more deeply integrated into
the MOM6 code, we can't easily wrap it with f2py (or
similar).  Instead, we pull the main logic of the coordinate out into
its own file (adaptive.F90) and interface that to Python with a custom
Cython file.  We pull almost the verbatim equation of state code out
of MOM6 to provide density derivative calculation routines.