Change wrapper to remap an entire array
Because we pass arrays as pointers to the Fortran routines, the data
of interest must be contiguous in memory.  This isn't the case for
k-indexed data, so we must introduce a copy to a contiguous array at
some point.  Because the use-case of the remapping code is almost
exclusively to remap the entire state of an array, we perform this at
the wrapper level, simplifying the Python interface to the remapping.