Make major changes to adaptive coordinate
Following some development work, we have significantly changed the
adaptive coordinate.  We have reframed the problem in terms of slope
tendencies, which allows the use of a direct physical limiter.  The
implementation isn't complete (only works for single dimension
slices), but it has both the adaptive and pressure smoothing terms.

The adaptive wrapper has been changed somewhat to return some
intermediate arrays for inspection, and the remapping wrapper exposes
a function to remap an entire array on the Cython side, hopefully
avoiding the expense of repeated calls between Python and
Cython/Fortran.